Pushing fixed muller error analysis
Sequence of events:
June 1st--Turns out that although k_var values were computed, that piece of code came before the calculation of the committor, and because of a return statement the committor was never computed so the error data was all zeroes.
This is why the "fixed bug in muller code" commit was pushed
Aug 2: We incorporated this data into the error analysis notebook and presented error plots in red white and black. note that these plots were used in the presentation on june 3rd.